+++ backend/server.js
@@ -43,5 +43,5 @@ app.use("/api/collector", require("./routes/collector"));
 app.use("/api/task-handler", require("./routes/bottle"));
 
 const PORT = process.env.PORT || 5000;
-app.listen(PORT, '0.0.0.0', () => console.log(`Server running on http://192.168.50.38:${PORT}`)); 
+app.listen(PORT, '0.0.0.0', () => console.log(`Server running on http://192.168.244.38:${PORT}`));
