+++ frontend_mobile/tsconfig.json
@@ -1,5 +1,4 @@
 {
-  "extends": "expo/tsconfig.base",
   "compilerOptions": {
     "strict": true,
     "paths": {
@@ -13,6 +12,8 @@
     "**/*.tsx",
     ".expo/types/**/*.ts",
     "expo-env.d.ts",
-    "app/index.js"
-, "app/login.js"  ]
+    "app/index.js",
+    "app/login.js"
+  ],
+  "extends": "expo/tsconfig.base"
 }
