+++ mobile_frontend/.env
@@ -1,1 +1,1 @@
-API_URL=http://192.168.50.38:5000/api
+EXPO_PUBLIC_API_URL=http://192.168.244.38:5000/api
