--- mobile_frontend/app/.env
@@ -1,1 +1,0 @@
-API_URL=http://192.168.65.221:5000/api
