+++ mobile_frontend/babel.config.js
@@ -1,14 +1,7 @@
-module.exports = function(api) {
+module.exports = function (api) {
   api.cache(true);
   return {
-    presets: ['babel-preset-expo'], 
-    plugins: [
-      ['module:react-native-dotenv', {
-        moduleName: '@env',
-        path: '.env',
-        safe: false,
-        allowUndefined: true,
-      }]
-    ],
+    presets: ['babel-preset-expo'],
+    
   };
 };
