+++ mobile_frontend/package.json
@@ -8,78 +8,50 @@
     "android": "expo start --android",
     "ios": "expo start --ios",
     "web": "expo start --web",
-    "test": "jest --watchAll",
     "lint": "expo lint"
   },
-  "jest": {
-    "preset": "jest-expo"
-  },
   "dependencies": {
-    "@expo/vector-icons": "^14.0.2",
-    "@react-google-maps/api": "^2.20.6",
-    "@react-native-async-storage/async-storage": "^1.23.1",
-    "@react-native-community/masked-view": "^0.1.11",
-    "@react-native-picker/picker": "^2.9.0",
-    "@react-navigation/bottom-tabs": "^7.2.0",
-    "@react-navigation/native": "^7.0.14",
-    "@react-navigation/stack": "^7.1.1",
-    "axios": "^1.7.9",
-    "expo": "^52.0.46",
-    "expo-barcode-scanner": "^13.0.1",
-    "expo-blur": "~14.0.3",
-    "expo-camera": "~16.0.18",
-    "expo-constants": "~17.0.6",
-    "expo-file-system": "~18.0.12",
-    "expo-font": "~13.0.3",
-    "expo-haptics": "~14.0.1",
-    "expo-linking": "~7.0.5",
-    "expo-media-library": "~17.0.6",
-    "expo-router": "~4.0.20",
-    "expo-sharing": "~13.0.1",
-    "expo-splash-screen": "^0.29.24",
-    "expo-status-bar": "~2.0.1",
-    "expo-symbols": "~0.2.2",
-    "expo-system-ui": "~4.0.9",
-    "expo-web-browser": "~14.0.2",
-    "file-saver": "^2.0.5",
-    "jsqr": "^1.4.0",
-    "patch-package": "^8.0.0",
-    "postinstall-postinstall": "^2.1.0",
-    "react": "18.3.1",
-    "react-dom": "18.3.1",
-    "react-google-autocomplete": "^2.7.4",
-    "react-native": "0.76.9",
-    "react-native-dotenv": "^3.4.11",
-    "react-native-fs": "^2.20.0",
-    "react-native-gesture-handler": "^2.20.2",
-    "react-native-maps": "^1.18.0",
-    "react-native-paper": "^5.13.2",
-    "react-native-permissions": "^5.3.0",
-    "react-native-qrcode-svg": "^6.3.15",
-    "react-native-reanimated": "^3.16.1",
-    "react-native-safe-area-context": "4.12.0",
-    "react-native-screens": "^4.4.0",
-    "react-native-svg": "^15.8.0",
-    "react-native-view-shot": "^4.0.3",
-    "react-native-vision-camera": "^4.6.4",
-    "react-native-web": "^0.19.13",
-    "react-native-webview": "^13.12.5",
-    "react-native-zip-archive": "^7.0.1",
-    "react-qr-code": "^2.0.15",
-    "react-webcam": "^7.2.0",
-    "vision-camera-code-scanner": "^0.2.0"
+    "@expo/vector-icons": "^14.1.0",
+    "@react-native-async-storage/async-storage": "2.1.2",
+    "@react-native-picker/picker": "2.11.0",
+    "@react-navigation/bottom-tabs": "^7.3.10",
+    "@react-navigation/elements": "^2.3.8",
+    "@react-navigation/native": "^7.1.6",
+    "axios": "^1.9.0",
+    "expo": "~53.0.8",
+    "expo-blur": "~14.1.4",
+    "expo-camera": "~16.1.6",
+    "expo-constants": "~17.1.6",
+    "expo-font": "~13.3.1",
+    "expo-haptics": "~14.1.4",
+    "expo-image": "~2.1.7",
+    "expo-linking": "~7.1.4",
+    "expo-location": "~18.1.4",
+    "expo-media-library": "~17.1.6",
+    "expo-router": "~5.0.6",
+    "expo-splash-screen": "~0.30.8",
+    "expo-status-bar": "~2.2.3",
+    "expo-symbols": "~0.4.4",
+    "expo-system-ui": "~5.0.7",
+    "expo-web-browser": "~14.1.6",
+    "react": "19.0.0",
+    "react-dom": "19.0.0",
+    "react-native": "0.79.2",
+    "react-native-gesture-handler": "~2.24.0",
+    "react-native-maps": "1.20.1",
+    "react-native-paper": "^5.14.0",
+    "react-native-reanimated": "~3.17.4",
+    "react-native-safe-area-context": "5.4.0",
+    "react-native-screens": "~4.10.0",
+    "react-native-web": "^0.20.0",
+    "react-native-webview": "13.13.5"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
-    "@react-native-community/cli": "latest",
-    "@types/jest": "^29.5.12",
-    "@types/react": "~18.3.12",
-    "@types/react-test-renderer": "^18.3.0",
-    "jest": "^29.2.1",
-    "jest-expo": "~52.0.6",
-    "metro-react-native-babel-preset": "^0.77.0",
-    "react-test-renderer": "18.3.1",
-    "typescript": "^5.3.3"
+    "@types/react": "~19.0.10",
+    "eslint": "^9.25.0",
+    "eslint-config-expo": "~9.2.0",
+    "typescript": "~5.8.3"
   },
   "private": true
 }
