--- mobile_frontend/services/api.js
@@ -1,21 +1,0 @@
-/*import axios from "axios";
-
-const API_URL = "http://localhost:5000/api"; // Replace with your backend IP
-
-export const login = async (email, password) => {
-  try {
-    const response = await axios.post(`${API_URL}/auth/login`, { email, password });
-    return response.data;
-  } catch (error) {
-    throw error.response?.data?.message || "Login failed";
-  }
-};
-
-export const register = async (userData) => {
-  try {
-    const response = await axios.post(`${API_URL}/auth/register`, userData);
-    return response.data;
-  } catch (error) {
-    throw error.response?.data?.message || "Registration failed";
-  }
-};*/
