+++ mobile_frontend/tsconfig.json
@@ -1,4 +1,5 @@
 {
+  "extends": "expo/tsconfig.base",
   "compilerOptions": {
     "strict": true,
     "paths": {
@@ -11,8 +12,6 @@
     "**/*.ts",
     "**/*.tsx",
     ".expo/types/**/*.ts",
-    "expo-env.d.ts",
-    "app/oneindex.js"
-, "app/index.js"  ],
-  "extends": "expo/tsconfig.base"
+    "expo-env.d.ts"
+, "app/_layout.js", "app/index.js"  ]
 }
